requirements.txt
loguru==0.7.3
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
streamlit==1.42.0
SQLAlchemy==2.0.38
numpy==2.0.2